feat: 모든 서비스 Data Jpa Repository 적용

+++ gift-card-service/src/main/java/com/hyoguoo/giftcardservice/giftcard/application/dto/command/FindMyGiftCardCursorCommand.java
@@ -9,5 +9,5 @@ public class FindMyGiftCardCursorCommand {
 
     private final Long userId;
     private final Long cursor;
-    private final Long size;
+    private final Integer size;
 }
